New branch!
As it would turn out I will probably need to do a full re-write of the code so that it works with structures instead of many arrays.

Consider the older struct-development branch dead now, all structure development will be done in this branch because things are abot to get ugly.